requirements.txt
streamlit
supabase
sentence-transformers
python-dotenv
requests
textblob